pyproject.toml
[project]
name = "home-ui"
version = "0.1.0"
description = "A lightweight homepage dashboard for self-hosted services"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "python-multipart>=0.0.6",
    "jinja2>=3.1.2",
    "pyyaml>=6.0.1",
    "pydantic-settings>=2.0.3",
    "itsdangerous>=2.1.2",
]
